+++ transactionhelper/app.py
@@ -1,4 +1,0 @@
-
-from . import TransactionHelper as RealTransactionHelper
-
-TransactionHelper = RealTransactionHelper